NABTO-1332: shared tcp and uart functions seperated into seperate files
