+++ src/modules/multi_tunnel/README.md
@@ -1,8 +1,8 @@
-Tunnel module
-=============
+Multi tunnel module
+===================
 
 The tunnelling module is used such that the device can be used as a
-transparent uart tunnel together with the nabto client.
+transparent uart, tcp, or echo tunnel together with the nabto client.
 
 Test Level
 ----------
