+++ src/modules/tunnel_common/README.md
@@ -1,0 +1,11 @@
+Common tunnel module
+====================
+
+The common tunnel module implements functions used when creating TCP,
+ UART, or Multi tunnels.
+ 
+
+Test Level
+----------
+
+This code is tested for each release.
