+++ src/modules/tunnel_common/tunnel_uart.c
@@ -1,0 +1,314 @@
+
+#include <unabto/unabto_stream.h>
+#include <unabto/unabto_memory.h>
+#include <errno.h>
+#include <fcntl.h>
+#include <modules/network/epoll/unabto_epoll.h>
+
+#include <termios.h>
+
+#include <modules/tunnel_common/tunnel_common.h>
+#include <modules/tunnel_common/tunnel_uart.h>
+
+#if NABTO_ENABLE_EPOLL
+#include <sys/epoll.h>
+#endif
+
+#define MAX(a,b) (((a)>(b))?(a):(b))
+#define MIN(x, y) (((x) < (y)) ? (x) : (y))
+
+extern void close_stream_reader(tunnel* tunnel);
+
+bool open_uart(tunnel* tunnel) {
+    struct termios tios;
+    
+    tunnel->tunnel_type_vars.uart.fd = open(tunnel->staticMemory->stmu.uart_sm.deviceName, O_RDWR | O_NOCTTY | O_NONBLOCK);
+    if (tunnel->tunnel_type_vars.uart.fd == -1) {
+        NABTO_LOG_ERROR(("cannot open uart %s", tunnel->staticMemory->stmu.uart_sm.deviceName));
+        return false;
+    }
+
+    memset(&tios, 0, sizeof(tios));
+    tios.c_iflag = IGNBRK | IGNCR;
+    tios.c_iflag |= ((tcflag_t) INPCK);
+    tios.c_oflag &= ~OPOST;
+    tios.c_lflag &= ~ICANON;
+    cfmakeraw(&tios); // :...
+    tios.c_cc[VMIN] = 0;
+    tios.c_cc[VTIME] = 0;
+    tios.c_cflag = CLOCAL | CREAD;
+        
+    
+    /* switch(parity) */
+    /* { */
+	/* 	case UART_PARITY_NONE: */
+	/* 		break; */
+			
+	/* 	case UART_PARITY_EVEN: */
+	/* 		tios.c_cflag |= ((tcflag_t) PARENB); */
+	/* 		break; */
+			
+	/* 	case UART_PARITY_ODD: */
+	/* 		tios.c_cflag |= ((tcflag_t) PARENB | PARODD); */
+	/* 		break; */
+			
+	/* 	// case UART_PARITY_MARK: */
+	/* 		// tios.c_cflag |= ((tcflag_t) PARENB | CMSPAR | PARODD); */
+	/* 		// break; */
+			
+	/* 	// case UART_PARITY_SPACE: */
+	/* 		// tios.c_cflag |= ((tcflag_t) PARENB | CMSPAR); */
+	/* 		// break; */
+
+	/* 	default: */
+	/* 		NABTO_LOG_FATAL(("Invalid number of databits for UART!")); */
+    /* } */
+
+    /* switch(stopbits) */
+    /* { */
+	/* 	case UART_STOPBITS_ONE: */
+	/* 		break; */
+
+	/* 	case UART_STOPBITS_TWO: */
+	/* 		tios.c_cflag |= ((tcflag_t) CSTOPB); */
+	/* 		break; */
+
+	/* 	default: */
+	/* 		NABTO_LOG_FATAL(("Invalid number of stopbits for UART!")); */
+    /* } */
+	
+	/* switch(databits) */
+	/* { */
+	/* 	case 5: */
+	/* 		tios.c_cflag |= ((tcflag_t) CS5); */
+	/* 		break; */
+	/* 	case 6: */
+	/* 		tios.c_cflag |= ((tcflag_t) CS6); */
+	/* 		break; */
+	/* 	case 7: */
+	/* 		tios.c_cflag |= ((tcflag_t) CS7); */
+	/* 		break; */
+	/* 	case 8: */
+	/* 		tios.c_cflag |= ((tcflag_t) CS8); */
+	/* 		break; */
+	/* 	default: */
+	/* 		NABTO_LOG_FATAL(("Invalid number of databits specified for UART '%s'.", _name)); */
+	/* 		return; */
+	/* } */
+    
+    /* switch(baudrate) */
+	/* { */
+	/* 	case 300: */
+	/* 		cfsetispeed(&tios, B300); */
+	/* 		cfsetospeed(&tios, B300); */
+	/* 	case 600: */
+	/* 		cfsetispeed(&tios, B600); */
+	/* 		cfsetospeed(&tios, B600); */
+	/* 	case 1200: */
+	/* 		cfsetispeed(&tios, B1200); */
+	/* 		cfsetospeed(&tios, B1200); */
+	/* 	case 2400: */
+	/* 		cfsetispeed(&tios, B2400); */
+	/* 		cfsetospeed(&tios, B2400); */
+	/* 		break; */
+	/* 	case 4800: */
+	/* 		cfsetispeed(&tios, B4800); */
+	/* 		cfsetospeed(&tios, B4800); */
+	/* 		break; */
+	/* 	case 9600: */
+	/* 		cfsetispeed(&tios, B9600); */
+	/* 		cfsetospeed(&tios, B9600); */
+	/* 		break; */
+	/* 	case 19200: */
+	/* 		cfsetispeed(&tios, B19200); */
+	/* 		cfsetospeed(&tios, B19200); */
+	/* 		break; */
+	/* 	case 38400: */
+	/* 		cfsetispeed(&tios, B38400); */
+	/* 		cfsetospeed(&tios, B38400); */
+	/* 		break; */
+	/* 	case 57600: */
+	/* 		cfsetispeed(&tios, B57600); */
+	/* 		cfsetospeed(&tios, B57600); */
+	/* 		break; */
+	/* 	case 115200: */
+	/* 		cfsetispeed(&tios, B115200); */
+	/* 		cfsetospeed(&tios, B115200); */
+	/* 		break; */
+	/* 	default: */
+	/* 		NABTO_LOG_FATAL(("Invalid baudrate specified for UART '%s'.", _name)); */
+	/* 		return; */
+	/* } */
+
+    // defautlt settings for now
+    
+    tios.c_cflag |= ((tcflag_t) CS8);
+    if (-1 == cfsetispeed(&tios, B115200)) {
+        NABTO_LOG_ERROR(("unable to set baudrate"));
+    }
+    if (-1 == cfsetospeed(&tios, B115200)) {
+        NABTO_LOG_ERROR(("unable to set baudrate"));
+    }
+    
+    if (-1 == tcsetattr(tunnel->tunnel_type_vars.uart.fd, TCSANOW, &tios))
+	{
+		NABTO_LOG_ERROR(("Unable to configure UART '%s'. error: %s", tunnel->staticMemory->stmu.uart_sm.deviceName, strerror(errno)));
+        return false;
+	}
+    
+#if NABTO_ENABLE_EPOLL
+    {
+        struct epoll_event ev;
+        ev.events = EPOLLIN | EPOLLOUT | EPOLLET;
+        ev.data.ptr = tunnel;
+	tunnel->epollEventType = UNABTO_EPOLL_TYPE_UART_TUNNEL;
+        epoll_ctl(unabto_epoll_fd, EPOLL_CTL_ADD, tunnel->tunnel_type_vars.uart.fd, &ev);
+    }
+#endif
+
+    
+    tunnel->state = TS_FORWARD;
+    return true;
+}
+
+/**
+ * read from uart, write to unabto
+ */
+void uart_forward(tunnel* tunnel) {
+    while(true) {
+        unabto_stream_hint hint;
+        size_t canWriteToStreamBytes;
+
+        if (tunnel->extReadState == FS_CLOSING) {
+            break;
+        }
+        
+        canWriteToStreamBytes = unabto_stream_can_write(tunnel->stream, &hint);
+        if (hint != UNABTO_STREAM_HINT_OK) {
+            close_reader(tunnel);
+            break;
+        }
+        if (canWriteToStreamBytes == 0) {
+            tunnel->extReadState = FS_WRITE;
+            break;
+        } else {
+            tunnel->extReadState = FS_READ;
+        }
+        {
+            int readen;
+            uint8_t readBuffer[NABTO_MEMORY_STREAM_SEND_SEGMENT_SIZE];
+            size_t maxRead = MIN(NABTO_MEMORY_STREAM_SEND_SEGMENT_SIZE, canWriteToStreamBytes);
+
+            readen = read(tunnel->tunnel_type_vars.uart.fd, readBuffer, maxRead);
+            if (readen == 0) {
+                // no data available
+                break;
+            }
+            
+            if (readen < 0) {
+                if (errno == EAGAIN || errno == EWOULDBLOCK) {
+                    break;
+                } else {
+                    close_reader(tunnel);
+                    break;
+                }
+            }
+            
+            if (readen > 0) {
+                unabto_stream_hint hint;
+                size_t written = unabto_stream_write(tunnel->stream, readBuffer, readen, &hint);
+                // the earlier call to unabto_stream_can_write
+                // ensures we can write all the bytes to the
+                // stream in one write
+                if (hint != UNABTO_STREAM_HINT_OK) {
+                    NABTO_LOG_TRACE(("Can't write to stream"));
+                    close_reader(tunnel);
+                    break;
+                }
+                
+                if (written != readen) {
+                    // Invalid state
+                    NABTO_LOG_ERROR(("Impossible state! wanted to write %i, wrote %i, unabto_said it could write %i bytes", readen, written, canWriteToStreamBytes));
+                }
+            }
+        }
+    }
+}
+
+
+/**
+ * read from unabto, write to uart
+ */
+void unabto_forward_uart(tunnel* tunnel) {
+    if (tunnel->unabtoReadState == FS_WRITE) {
+        tunnel->unabtoReadState = FS_READ;
+    }
+    while(true) {
+        if (tunnel->unabtoReadState == FS_READ) {
+            const uint8_t* buf;
+            unabto_stream_hint hint;
+            size_t readen = unabto_stream_read(tunnel->stream, &buf, &hint);
+            if (hint != UNABTO_STREAM_HINT_OK) {
+                close_stream_reader(tunnel);
+                break;
+            } else {
+                if (readen == 0) {
+                    break;
+                } else {
+                    ssize_t written;
+                    NABTO_LOG_TRACE(("Write to uart stream %i", readen));
+                    written = write(tunnel->tunnel_type_vars.uart.fd, buf, readen);
+
+                    if (written > 0) {
+                        NABTO_LOG_TRACE(("Wrote to tcp stream %i", written));
+                        unabto_stream_ack(tunnel->stream, buf, written, &hint);
+                        if (hint != UNABTO_STREAM_HINT_OK) {
+                            close_stream_reader(tunnel);
+                            break;
+                        }
+                    } else if (written == 0) {
+                        tunnel->unabtoReadState = FS_WRITE;
+                        break;
+                    } else { // -1
+                        NABTO_LOG_TRACE(("Wrote to uart stream %i, status %s", written, strerror(errno)));
+
+                        if (errno == EAGAIN || errno == EWOULDBLOCK) {
+                            tunnel->unabtoReadState = FS_WRITE;
+                            break;
+                        } else {
+                            close_stream_reader(tunnel);
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+
+        if (tunnel->unabtoReadState == FS_WRITE) {
+            break;
+        }
+
+        if (tunnel->unabtoReadState == FS_CLOSING) {
+            break;
+        }
+    }
+}
+
+/**
+ * utility function for either sending +ok\n or -error message\n in
+ * case of either success or failure.  see spec in
+ * https://www.rfc-editor.org/rfc/rfc1078.txt instead of CRLF only LF
+ * is used.
+ */
+bool tunnel_send_init_message(tunnel* tunnel, const char* msg)
+{
+    unabto_stream_hint hint;
+    size_t written;
+    size_t writeLength = strlen(msg);
+    written = unabto_stream_write(tunnel->stream, (uint8_t*)msg, writeLength, &hint);
+    if (written != writeLength) {
+        NABTO_LOG_ERROR(("we should at a minimum be able to send this simple message, we will probably just goive up..."));
+        return false;
+    }
+    return true;
+}
